backend/requirements.txt
python-dotenv
fastapi
uvicorn
ollama
requests
pydantic
mistralai
openai
google-genai
anthropic
ftfy